Separate ngraphFrontend-related code from other
