+++ model-optimizer/automation/package_BOM.txt
@@ -911,6 +911,11 @@ mo/front/tf/partial_infer/__init__.py
 mo/front/tf/partial_infer/tf.py
 mo/front/tf/register_custom_ops.py
 mo/front/tf/replacement.py
+mo/front_ng/__init__.py
+mo/front_ng/extractor.py
+mo/front_ng/frontendmanager_wrapper.py
+mo/front_ng/pipeline.py
+mo/front_ng/serialize.py
 mo/graph/__init__.py
 mo/graph/connection.py
 mo/graph/graph.py
@@ -992,7 +997,6 @@ mo/utils/error.py
 mo/utils/extract_release_version.py
 mo/utils/find_ie_version.py
 mo/utils/find_inputs.py
-mo/utils/frontendmanager_wrapper.py
 mo/utils/get_ov_update_message.py
 mo/utils/graph.py
 mo/utils/guess_framework.py
