+++ model-optimizer/mo/front/extractor.py
@@ -451,34 +451,6 @@ def extract_node_attrs(graph: Graph, extractor: callable):
     return graph
 
 
-def decodeNameWithPort (graph: Graph, node_name: str):
-    """
-    Decode name with optional port specification w/o traversing all the nodes in the graph
-    :param graph:
-    :param node_name:
-    :return: decoded place in the graph
-    """
-    # Check exact match with one of the names in the graph first
-    inputModel = graph.graph['input_model']
-    #TODO: search for any name, not only a tensor
-    node = inputModel.getPlaceByTensorName(node_name)
-    if node:
-        return node
-    regexpPost = r'(.*)(:(\d+))'
-    matchPost = re.search(regexpPost, node_name)
-    nodePost = inputModel.getPlaceByTensorName(matchPost.group(1)) if matchPost else None
-    regexpPre = r'((\d+):)(.*)'
-    matchPre = re.search(regexpPre, node_name)
-    nodePre = inputModel.getPlaceByTensorName(matchPre.group(3)) if matchPost else None
-    if nodePost and nodePre:
-        raise Error('Name collision for {}'.format(node_name))
-    if nodePost:
-        return node.getOutputPort(int(matchPost.group(3)))
-    if nodePre:
-        return node.getInputPort(int(matchPre.group(1)))
-    raise Error('There is no node with name {}'.format(node_name))
-
-
 def get_node_id_with_ports(graph: Graph, node_name: str, skip_if_no_port=True):
     """
     Extracts port and node ID out of user provided name
@@ -580,38 +552,13 @@ def input_user_data_repack(graph: Graph, input_user_shapes: [None, list, dict, n
         'phase_train' : False
     }
     """
-    _input_shapes = []
-    if 'frontend' in graph.graph:
-        # New version of FrontEnd is activated
-        print("I'm HERE")
-        inputModel = graph.graph['input_model']
-        if isinstance(input_user_shapes, list) or isinstance(input_user_shapes, dict):
-            for input_name in input_user_shapes:
-                node = decodeNameWithPort(graph, input_name)
-                if node is None:
-                    raise Error('Cannot find location {} in the graph'.format(input_name))
-                shape = None if isinstance(input_user_shapes, list) else input_user_shapes[input_name]
-                if input_name in input_user_data_types and input_user_data_types[input_name] is not None:
-                    data_type = input_user_data_types[input_name]
-                    _input_shapes.append({'node': node, 'shape': shape, 'data_type': data_type})
-                else:
-                    _input_shapes.append({'node': node, 'shape': shape})
-        elif isinstance(input_user_shapes, np.ndarray):
-            model_inputs = inputModel.getInputs()
-            assert len(model_inputs) == 1
-            _input_shapes.append({'node': model_inputs[0], 'shape': input_user_shapes})
-        else:
-            assert input_user_shapes is None
-        return _input_shapes, dict()
-
-
     _input_shapes = defaultdict(list)
     _freeze_placeholder = dict()
     _freeze_new_placeholder = defaultdict(list)
 
     # freeze placeholder restructure
     # Replaces placeholder name with placeholder id. Raises if there is no placeholder with such ID
-    placeholders_ids = graph.get_input_ids()
+    placeholders_ids = graph.get_nodes_with_attributes(op='Parameter')
     if freeze_placeholder is None:
         _freeze_placeholder = None
     else:
@@ -713,19 +660,6 @@ def output_user_data_repack(graph: Graph, outputs: list):
             ]
     }
     """
-    if 'frontend' in graph.graph:
-        _outputs = []
-        # New version of FrontEnd is activated
-        print("I'm HERE - output")
-        print("Outputs is list or dict")
-        if outputs is not None and len(outputs) > 0:
-            for output in outputs:
-                node = decodeNameWithPort(graph, output)
-                if node is None:
-                    raise Error('Cannot find location {} in the graph'.format(output))
-                _outputs.append({'node': node})
-        return _outputs
-
     _outputs = defaultdict(list)
     if outputs is None:
         _outputs = None
@@ -1017,20 +951,6 @@ def add_input_ops(graph: Graph, user_defined_inputs: dict, before_infer: bool):
 
     For case with before_infer=False data nodes are added to this schemes.
     """
-
-    # Limited implementation for nGraph
-    # Support reshape only, no model cutting is available
-    #print('--- USER DEFINED INPUTS ---:' + str(user_defined_inputs))
-    if 'network' in graph.graph:
-        # warning connection through node name
-        #print(user_defined_inputs)
-        # temporary we handle only 0th index
-        if user_defined_inputs:
-            reshape_request = {k.get_friendly_name(): user_defined_inputs[k][0]['shape'] for k in user_defined_inputs}
-            #print('RESHAPE REQUEST: ' + str(reshape_request))
-            graph.graph['network'].reshape(reshape_request)
-        return
-
     inputs = []
     set_is_input(graph, graph.get_nodes_with_attributes(op='Parameter'), False)
     if user_defined_inputs is None:
